requirements.txt
flask
flask-cors
gunicorn
langchain==0.2.14
langchain-community==0.2.12
langchain-groq==0.1.9
langchain-huggingface==0.0.3
faiss-cpu
sentence-transformers
groq
python-dotenv
langchain-google-genai==1.0.10
markdown
